+++ src/Auth/Token.php
@@ -4,14 +4,15 @@ declare(strict_types=1);
 
 namespace BjTheCod3r\Spotify\Auth;
 
-use Illuminate\Support\Carbon;
+use Carbon\CarbonInterface;
+use Illuminate\Support\Facades\Date;
 
 final readonly class Token
 {
     public function __construct(
         public string $accessToken,
         public string $tokenType,
-        public Carbon $expiresAt,
+        public CarbonInterface $expiresAt,
     ) {
     }
 
@@ -45,7 +46,7 @@ final readonly class Token
         return new self(
             accessToken: $data['access_token'],
             tokenType: $data['token_type'],
-            expiresAt: Carbon::createFromTimestamp($data['expires_at']),
+            expiresAt: Date::createFromTimestamp($data['expires_at']),
         );
     }
 }