+++ src/Events/SpotifyTokenRefreshed.php
@@ -4,13 +4,13 @@ declare(strict_types=1);
 
 namespace BjTheCod3r\Spotify\Events;
 
-use Illuminate\Support\Carbon;
+use Carbon\CarbonInterface;
 
 final class SpotifyTokenRefreshed
 {
     public function __construct(
         public readonly int|string $userId,
-        public readonly Carbon $expiresAt,
+        public readonly CarbonInterface $expiresAt,
     ) {
     }
 }